refactor(data): finish catalog splits and add data NCL gate (#4072)
Move remaining tech categories into chunk parts, split combat/naming catalogs
and military GA params under the soft 500 NCL ceiling, and wire
repo.data_lib_file_size with SPEC/test coverage.
