+++ SPEC/program/repo-lint.md
@@ -75,6 +75,7 @@
 | `tool/check_map_gen_stage_protocol.dart` | Require the shared `MapGenStage` contract in `packages/colonizethis_map/lib/src/gen/map_gen_stage.dart` and `implements MapGenStage` (or the uniform `MapGenPass`) on each bound generator service family (`TileMapGenLandSeeds`, `TileMapGenLakesProvinces`, `ContinentJoinPass`, `TerrainJitterPass`, `SeaZoneSubdividePass`, `TileMapGenTerrainResource`) so pass orchestration shares a uniform params + grid-in/grid-out protocol, and require **at least 4** bound families to adopt the uniform `MapGenPass` entry point (raised from 3 after the former join-sea family was split into three standalone `MapGenPass` passes, removing the last `MapGenStage`-only exemption) (Refs #3459, #3574, #3588); rule `repo.map_gen_stage_protocol` |
 | `tool/check_map_no_partfile_classes.dart` | Two map structure gates share this script. `repo.map_no_partfile_classes` forbids any `part of` file under `packages/colonizethis_map/lib/src/**` from declaring a top-level non-private class. `repo.map_gen_no_new_partfiles` forbids **any** `part` / `part of` directive under the generation layer (`lib/src/gen/**`) and view layer (`lib/src/view/**`) with an **empty** allowlist after #4022 promoted `tile_map_generator_types.dart` to a standalone library (closing the temporary #3588 exception); rules `repo.map_no_partfile_classes`, `repo.map_gen_no_new_partfiles` |
 | `tool/check_map_lib_file_size.dart` | Enforce that every `.dart` file under the colonizethis_map generation layer (`lib/src/gen/**`), view layer (`lib/src/view/**`), and render layer (`lib/src/render/**`) is **500 non-comment lines or fewer** (fail when strictly greater), reusing `countNonCommentLinesFromSource`; the gate also fails when any scanned root is missing so it cannot silently rot if those layers move. Complements the repository-wide `repo.dart_file_non_comment_line_size` cap so the #3588 / #3846 part-file decomposition keeps each pass/concern small (Refs #3588, #3846); rule `repo.map_lib_file_size` |
+| `tool/check_data_lib_file_size.dart` | Enforce that every hand-written `.dart` file under `packages/colonizethis_data/lib/src/**` is **500 non-comment lines or fewer** (fail when strictly greater), reusing `countNonCommentLinesFromSource`; skips generated suffixes and `tech_effect_summary_embed.dart`. Complements the repository-wide 1000 NCL cap so the #4072 topic splits stay small (Refs #4072); rule `repo.data_lib_file_size` |
 | `tool/check_map_public_barrel_surface.dart` | Keep the public `colonizethis_map` barrel `packages/colonizethis_map/lib/colonizethis_map.dart` narrowed to intentional entry points: forbid any `export` of the internal-only generation primitives `src/gen/grid_voronoi.dart` (deterministicNoise / assignCellsToNearestSeed), `src/gen/topology_inference.dart` (inferTopologyFromTileMap), and `src/gen/tile_map_grid_graph.dart` (TileMapGridGraph). Same-package tests import these from `src/` directly; comment lines and `import` directives are ignored (Refs #3459 AC4); rule `repo.map_public_barrel_surface` |
 | `tool/check_map_region_data_access_central.dart` | Keep old/new-world region data selection centralized in `packages/colonizethis_map/lib/**`: forbid any inline `.oldWorld.provinces`, `.oldWorld.units`, `.newWorld.provinces`, or `.newWorld.units` region branch outside the canonical helper file `packages/colonizethis_map/lib/src/region_data_access.dart`. Every region-scoped provinces/units read in the view-building (orchestration, cell/unit) and ownership-overlay passes must resolve the region via `regionDataForMapRegionId(world, regionId)` so the old/new selection has one source of truth; the canonical file is exempt and comment lines are ignored (Refs #3459 AC3); rule `repo.map_region_data_access_central` |
 | `tool/check_map_region_dispatch_central.dart` | Keep old/new-world region *dispatch* centralized in `packages/colonizethis_map/lib/**`: forbid any inline `==` / `!=` comparison against `kRegionOldWorld` or `kRegionNewWorld` outside the canonical dispatch file `packages/colonizethis_map/lib/src/map_region_dispatch.dart`. Colour, capital-scope, port-tile bucketing, and view-data assembly paths must resolve old/new branching via `selectByMapRegionId(...)` / `selectByMapRegionIdOrNull(...)` so dispatch has one source of truth. Using the constants as map keys, list elements, named arguments, or string-interpolation operands is unaffected; the canonical file is exempt and comment lines are ignored (Refs #3574); rule `repo.map_region_dispatch_central` |
@@ -560,6 +561,8 @@ Do **not** add new top-level `tool/check_*.dart` **entrypoints** for CI without 
 - Given a generation-layer or view-layer file that declares a `part` or `part of` directive, when `dart run tool/ct_repo_lint.dart` runs rule `repo.map_gen_no_new_partfiles`, then the run fails listing the offending `path:line:column` and exits `1` (Refs #3588, #4022).
 - Given the generation-layer files under `packages/colonizethis_map/lib/src/gen/**`, view-layer files under `lib/src/view/**`, and render-layer files under `lib/src/render/**`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.map_lib_file_size`, then every scanned `.dart` file has at most 500 non-comment lines and the run exits `0` (Refs #3588, #3846).
 - Given a generation-layer, view-layer, or render-layer `.dart` file with strictly more than 500 non-comment lines, when `dart run tool/ct_repo_lint.dart` runs rule `repo.map_lib_file_size`, then the run fails listing the offending path with its non-comment line count and exits `1` (Refs #3588, #3846).
+- Given every hand-written `.dart` file under `packages/colonizethis_data/lib/src/**` is at or below 500 non-comment lines (excluding `tech_effect_summary_embed.dart`), when `dart run tool/ct_repo_lint.dart` runs rule `repo.data_lib_file_size`, then the rule passes and exits `0` (Refs #4072).
+- Given a hand-written `packages/colonizethis_data/lib/src/**` Dart file with strictly more than 500 non-comment lines, when `dart run tool/ct_repo_lint.dart` runs rule `repo.data_lib_file_size`, then the run fails naming that file and exits `1` (Refs #4072).
 - Given more than 4 files matching `packages/colonizethis_map/test/init_game_map_view_builder_*_test.dart`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.map_test_view_builder_file_count`, then the run fails listing each matching path and exits `1` (Refs #3846).
 - Given the map view-builder marker suites after #4022 AC5, when counting `init_game_map_view_builder_*_test.dart` files, then the count remains ≤4 and ownership/capital/port/warp marker coverage lives in `init_game_map_view_builder_markers_test.dart` while town/civilian marker coverage lives in `init_game_map_view_builder_town_civilian_markers_test.dart` (visibility/presence coverage is folded into `init_game_map_view_builder_region_cells_test.dart`) (Refs #4022).
 - Given `packages/colonizethis_map/test/tile_map_visualization_*`, `test/game_world_state_map_visualizer_test.dart`, or `test/multi_region_map_rendering_test.dart` with an unallowlisted inline `\bTopologyNode\(`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.map_test_shared_topology_fixtures`, then the run fails listing each offending path and exits `1` (Refs #3846).
