+++ packages/colonizethis_data/lib/src/combat_regiment_catalog.dart
@@ -1,0 +1,296 @@
+part of 'combat_config.dart';
+
+/// Full regiment table: 29 types across 8 categories and 4 eras.
+/// SPEC/game/military-units.md.
+const List<RegimentStats> regimentCatalog = [
+  RegimentStats(
+    id: 'peasant_levies',
+    fpn: 0,
+    fpm: 3,
+    rng: 1,
+    def: 3,
+    mvr: 3,
+    category: RegimentCategory.lightInfantry,
+    era: 1,
+  ),
+  RegimentStats(
+    id: 'pikemen',
+    fpn: 0,
+    fpm: 5,
+    rng: 1,
+    def: 5,
+    mvr: 3,
+    category: RegimentCategory.regularInfantry,
+    era: 1,
+  ),
+  RegimentStats(
+    id: 'arquebusiers',
+    fpn: 5,
+    fpm: 1,
+    rng: 3,
+    def: 3,
+    mvr: 2,
+    category: RegimentCategory.heavyInfantry,
+    era: 1,
+  ),
+  RegimentStats(
+    id: 'bowmen',
+    fpn: 3,
+    fpm: 1,
+    rng: 4,
+    def: 2,
+    mvr: 3,
+    category: RegimentCategory.bowmen,
+    era: 1,
+  ),
+  RegimentStats(
+    id: 'squires',
+    fpn: 0,
+    fpm: 4,
+    rng: 1,
+    def: 4,
+    mvr: 6,
+    category: RegimentCategory.lightCavalry,
+    era: 1,
+  ),
+  RegimentStats(
+    id: 'knights',
+    fpn: 0,
+    fpm: 6,
+    rng: 1,
+    def: 6,
+    mvr: 4,
+    category: RegimentCategory.spearCavalry,
+    era: 1,
+  ),
+  RegimentStats(
+    id: 'culverin',
+    fpn: 8,
+    fpm: 1,
+    rng: 5,
+    def: 2,
+    mvr: 2,
+    category: RegimentCategory.heavyArtillery,
+    era: 1,
+  ),
+  RegimentStats(
+    id: 'calivermen',
+    fpn: 3,
+    fpm: 2,
+    rng: 5,
+    def: 5,
+    mvr: 4,
+    category: RegimentCategory.lightInfantry,
+    era: 2,
+  ),
+  RegimentStats(
+    id: 'halberdiers',
+    fpn: 0,
+    fpm: 7,
+    rng: 1,
+    def: 6,
+    mvr: 4,
+    category: RegimentCategory.regularInfantry,
+    era: 2,
+  ),
+  RegimentStats(
+    id: 'musketeers',
+    fpn: 7,
+    fpm: 2,
+    rng: 4,
+    def: 4,
+    mvr: 3,
+    category: RegimentCategory.heavyInfantry,
+    era: 2,
+  ),
+  RegimentStats(
+    id: 'cossacks',
+    fpn: 0,
+    fpm: 5,
+    rng: 1,
+    def: 5,
+    mvr: 8,
+    category: RegimentCategory.lightCavalry,
+    era: 2,
+  ),
+  RegimentStats(
+    id: 'lancers',
+    fpn: 0,
+    fpm: 8,
+    rng: 1,
+    def: 5,
+    mvr: 6,
+    category: RegimentCategory.spearCavalry,
+    era: 2,
+  ),
+  RegimentStats(
+    id: 'harquebusiers',
+    fpn: 2,
+    fpm: 6,
+    rng: 3,
+    def: 5,
+    mvr: 6,
+    category: RegimentCategory.heavyCavalry,
+    era: 2,
+  ),
+  RegimentStats(
+    id: 'horse_artillery',
+    fpn: 5,
+    fpm: 2,
+    rng: 7,
+    def: 2,
+    mvr: 3,
+    category: RegimentCategory.lightArtillery,
+    era: 2,
+  ),
+  RegimentStats(
+    id: 'royal_artillery',
+    fpn: 9,
+    fpm: 2,
+    rng: 8,
+    def: 2,
+    mvr: 2,
+    category: RegimentCategory.heavyArtillery,
+    era: 2,
+  ),
+  RegimentStats(
+    id: 'skirmishers',
+    fpn: 4,
+    fpm: 3,
+    rng: 5,
+    def: 6,
+    mvr: 6,
+    category: RegimentCategory.lightInfantry,
+    era: 3,
+  ),
+  RegimentStats(
+    id: 'regulars',
+    fpn: 7,
+    fpm: 7,
+    rng: 5,
+    def: 5,
+    mvr: 4,
+    category: RegimentCategory.regularInfantry,
+    era: 3,
+  ),
+  RegimentStats(
+    id: 'grenadiers',
+    fpn: 10,
+    fpm: 8,
+    rng: 5,
+    def: 5,
+    mvr: 4,
+    category: RegimentCategory.heavyInfantry,
+    era: 3,
+  ),
+  RegimentStats(
+    id: 'hussars',
+    fpn: 2,
+    fpm: 8,
+    rng: 3,
+    def: 6,
+    mvr: 11,
+    category: RegimentCategory.lightCavalry,
+    era: 3,
+  ),
+  RegimentStats(
+    id: 'cuirassiers',
+    fpn: 5,
+    fpm: 13,
+    rng: 3,
+    def: 5,
+    mvr: 9,
+    category: RegimentCategory.heavyCavalry,
+    era: 3,
+  ),
+  RegimentStats(
+    id: 'light_artillery',
+    fpn: 8,
+    fpm: 3,
+    rng: 9,
+    def: 3,
+    mvr: 4,
+    category: RegimentCategory.lightArtillery,
+    era: 3,
+  ),
+  RegimentStats(
+    id: 'heavy_artillery',
+    fpn: 13,
+    fpm: 2,
+    rng: 10,
+    def: 2,
+    mvr: 3,
+    category: RegimentCategory.heavyArtillery,
+    era: 3,
+  ),
+  RegimentStats(
+    id: 'sharpshooters',
+    fpn: 5,
+    fpm: 4,
+    rng: 7,
+    def: 7,
+    mvr: 7,
+    category: RegimentCategory.lightInfantry,
+    era: 4,
+  ),
+  RegimentStats(
+    id: 'rifle_infantry',
+    fpn: 9,
+    fpm: 9,
+    rng: 6,
+    def: 6,
+    mvr: 4,
+    category: RegimentCategory.regularInfantry,
+    era: 4,
+  ),
+  RegimentStats(
+    id: 'guards',
+    fpn: 12,
+    fpm: 10,
+    rng: 6,
+    def: 6,
+    mvr: 4,
+    category: RegimentCategory.heavyInfantry,
+    era: 4,
+  ),
+  RegimentStats(
+    id: 'scouts',
+    fpn: 5,
+    fpm: 11,
+    rng: 5,
+    def: 6,
+    mvr: 11,
+    category: RegimentCategory.lightCavalry,
+    era: 4,
+  ),
+  RegimentStats(
+    id: 'carbine_cavalry',
+    fpn: 7,
+    fpm: 17,
+    rng: 5,
+    def: 5,
+    mvr: 9,
+    category: RegimentCategory.heavyCavalry,
+    era: 4,
+  ),
+  RegimentStats(
+    id: 'field_artillery',
+    fpn: 10,
+    fpm: 3,
+    rng: 11,
+    def: 4,
+    mvr: 5,
+    category: RegimentCategory.lightArtillery,
+    era: 4,
+  ),
+  RegimentStats(
+    id: 'siege_guns',
+    fpn: 17,
+    fpm: 2,
+    rng: 12,
+    def: 3,
+    mvr: 3,
+    category: RegimentCategory.heavyArtillery,
+    era: 4,
+  ),
+];
