+++ tool/ct_repo_lint_manifest.yaml
@@ -1079,6 +1079,14 @@ rules:
     script: tool/check_models_file_size.dart
     pr_incremental: true
 
+  - rule_id: repo.data_lib_file_size
+    group: structure
+    title: colonizethis_data lib/src files must stay at or below 500 non-comment lines (Refs #4072)
+    spec: SPEC/program/dart-file-non-comment-line-size.md
+    runner: dart
+    script: tool/check_data_lib_file_size.dart
+    pr_incremental: true
+
   - rule_id: repo.canonical_province_tile_keys
     group: game_invariants
     title: Province WorkOrder targetTileKey literals use full tile shape
